// types/generated/contentful.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IBlogPostFields {
  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Hero Image */
  heroImage: Asset;

  /** Description */
  description: string;

  /** Body */
  body: string;

  /** Author */
  author?: IPerson | undefined;

  /** Publish Date */
  publishDate: string;

  /** Tags */
  tags?: ("general" | "javascript" | "static-sites")[] | undefined;
}

export interface IBlogPost extends Entry<IBlogPostFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "blogPost";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IFeaturesFields {
  /** Heading */
  heading: string;

  /** Description */
  description: string;

  /** Image */
  image?: Asset | undefined;

  /** Image Left */
  imageLeft?: boolean | undefined;

  /** Bottom Spacer */
  bottomSpacer?: boolean | undefined;

  /** Background Color */
  backgroundColor?: string | undefined;

  /** Project */
  project?: IProject | undefined;
}

export interface IFeatures extends Entry<IFeaturesFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "features";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPageFields {
  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Heading */
  heading?: string | undefined;

  /** Image */
  image?: Asset | undefined;

  /** Description */
  description?: string | undefined;

  /** Body */
  body?: string | undefined;
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "page";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPersonFields {
  /** Name */
  name: string;

  /** Title */
  title: string;

  /** Company */
  company: string;

  /** Short Bio */
  shortBio: string;

  /** Email */
  email?: string | undefined;

  /** Phone */
  phone?: string | undefined;

  /** Facebook */
  facebook?: string | undefined;

  /** Twitter */
  twitter?: string | undefined;

  /** Github */
  github?: string | undefined;

  /** Image */
  image?: Asset | undefined;
}

export interface IPerson extends Entry<IPersonFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "person";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IProjectFields {
  /** Title */
  title: string;

  /** Type */
  type?: string[] | undefined;

  /** Slug */
  slug: string;

  /** Description */
  description?: string | undefined;

  /** Image */
  image: Asset;

  /** Stack */
  stack: string[];

  /** Details */
  details?: string[] | undefined;

  /** Mozaic */
  mozaic?: Asset[] | undefined;

  /** Features */
  features?: IFeatures[] | undefined;

  /** Numbers */
  numbers?: Record<string, any> | undefined;

  /** Challenges */
  challenges: string[];
}

export interface IProject extends Entry<IProjectFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "project";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE =
  | "blogPost"
  | "features"
  | "page"
  | "person"
  | "project";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";